+++ src/openzaak/js/components/admin/applicatie-autorisaties/autorisatie-formset.js
@@ -3,7 +3,20 @@ import PropTypes from "prop-types";
 
 import { AutorisatieForm } from './autorisatie-form';
 import { ManagementForm} from "../../../forms/management-form";
-import { AddRow} from "../../../forms/add-row";
+
+
+const AddRow = ({ className="add-row", onAdd, children }) => {
+    return (
+        <div className={className}>
+            <a href="#" onClick={onAdd}>{ children }</a>
+        </div>
+    );
+};
+
+AddRow.propTypes = {
+    className: PropTypes.string,
+    onAdd: PropTypes.func.isRequired,
+};
 
 
 const AutorisatieFormSet = (props) => {
