+++ src/openzaak/sass/admin/components/_all.scss
@@ -4,6 +4,7 @@
 @import "autorisatie-formset";
 @import "autorisatie-form";
 @import "catalogus-options";
+@import "external-formset";
 @import "external-form";
 
 @import "errorlist";
