+++ src/openzaak/sass/admin/components/_external-formset.scss
@@ -1,0 +1,17 @@
+.external-formset {
+
+  &__add-row a {
+      background: url('../../admin/img/icon-addlink.svg') 0 1px no-repeat;
+      padding-left: 16px;
+      font-size: 12px;
+  }
+
+  &__col--wide {
+      width: 25em;
+  }
+
+   &__col {
+      width: 18em;
+  }
+
+}
